refactor: move recovery middleware to middleware package
