+++ bastion.go
@@ -123,7 +123,7 @@ func initialize(app *Bastion) {
 	app.APIRouter = chi.NewRouter()
 	api500Err := errors.New(app.Options.API500ErrMessage)
 	app.APIRouter.Use(middleware.APIErrHandler(api500Err, app.Logger))
-	app.APIRouter.Use(Recovery(app.Logger))
+	app.APIRouter.Use(middleware.Recovery(app.Logger))
 	app.APIRouter.Use(LoggerRequest(app.Options)...)
 	app.r.Mount(app.Options.APIBasepath, app.APIRouter)
